Modify MarchingSquare and ScanLineRasterize in order to work with multi value

+++ hortonmachine/src/main/java/org/jgrasstools/hortonmachine/modules/basin/basinshape/BasinShape.java
@@ -241,7 +241,7 @@ public class BasinShape extends JGTModel {
                     squares.inGeodata = inBasins;
                     squares.pm = pm;
                     squares.doReset = true;
-                    squares.pValue = num;
+                    squares.pValue = (double) num;
                     squares.process();
                 } catch (Exception e) {
                     pm.errorMessage(e.getLocalizedMessage());
